move requests into submodule
