+++ mycloud/mycloudapi/__init__.py
@@ -1,9 +1,2 @@
-from mycloud.mycloudapi.metadata_request import MetadataRequest
-from mycloud.mycloudapi.object_request import GetObjectRequest, PutObjectRequest, DeleteObjectRequest
 from mycloud.mycloudapi.object_resource_builder import ObjectResourceBuilder
-from mycloud.mycloudapi.request import MyCloudRequest
 from mycloud.mycloudapi.request_executor import MyCloudRequestExecutor
-from mycloud.mycloudapi.change_request import ChangeRequest
-from mycloud.mycloudapi.usage_request import UsageRequest
-from mycloud.mycloudapi.rename_request import RenameRequest
-from mycloud.mycloudapi.directory_list_request import DirectoryListRequest, ListType
