+++ mycloud/mycloudapi/request_executor.py
@@ -4,9 +4,7 @@ import requests
 from requests.models import PreparedRequest
 from mycloud.logger import add_request_count, save_files
 from mycloud.mycloudapi.auth import MyCloudAuthenticator, AuthMode
-from mycloud.mycloudapi import MyCloudRequest
-from mycloud.mycloudapi.request import ContentType
-from mycloud.mycloudapi.request import Method
+from mycloud.mycloudapi.requests import Method, ContentType, MyCloudRequest
 from mycloud.constants import WAIT_TIME_MULTIPLIER, RESET_SESSION_EVERY
