+++ mycloud/mycloudapi/requests/drive/change_request.py
@@ -1,7 +1,7 @@
 import json
 import time
 from mycloud.mycloudapi.helper import get_object_id, raise_if_invalid_cloud_path
-from mycloud.mycloudapi.request import MyCloudRequest, Method
+from mycloud.mycloudapi.requests import MyCloudRequest, Method
 
 
 REQUEST_URL = 'https://storage.prod.mdl.swisscom.ch/changes/files/?p={}&$limit={}&nocache={}'
