+++ mycloud/mycloudapi/requests/drive/directory_list_request.py
@@ -2,7 +2,7 @@ import json
 from enum import Enum
 from time import time
 from mycloud.mycloudapi.helper import get_object_id, raise_if_invalid_cloud_path
-from mycloud.mycloudapi.request import MyCloudRequest, Method
+from mycloud.mycloudapi.requests import MyCloudRequest, Method
 
 
 REQUEST_URL = 'https://storage.prod.mdl.swisscom.ch/sync/list?p={}&$type={}&nocache={}'
