+++ mycloud/mycloudapi/requests/drive/metadata_request.py
@@ -1,6 +1,6 @@
 import json
 from mycloud.mycloudapi.helper import get_object_id, raise_if_invalid_cloud_path
-from mycloud.mycloudapi.request import MyCloudRequest, Method
+from mycloud.mycloudapi.requests import MyCloudRequest, Method
 
 
 REQUEST_URL = 'https://storage.prod.mdl.swisscom.ch/metadata?p='
