+++ mycloud/mycloudapi/requests/drive/rename_request.py
@@ -1,7 +1,7 @@
 import json
-from mycloud.mycloudapi.request import MyCloudRequest, Method
+from mycloud.mycloudapi.requests import MyCloudRequest, Method
 from mycloud.mycloudapi.object_resource_builder import ObjectResourceBuilder
-from mycloud.common import get_string_generator
+from mycloud.common.functions import get_string_generator
 
 
 REQUEST_URL = 'https://storage.prod.mdl.swisscom.ch/commands/rename'
